feat: Add Form class and signForm method to Bureaucrat class
This commit introduces a new Form class in Form.hpp and Form.cpp. It also adds a signForm method to the Bureaucrat class in Bureaucrat.hpp and Bureaucrat.cpp. The signForm method allows a Bureaucrat to sign a Form. If the Bureaucrat's grade is too low to sign the Form, an exception is thrown and caught within the method, and an appropriate error message is printed to the console.